Cleanup gem structure.
