+++ example/thin_rack_app.rb
@@ -3,19 +3,19 @@
 #  examples/thin_rack_app.rb
 #
 #  Just a simple example of running with thin.
-require 'lib/metrics_reporter'
+require 'lib/rack_middleware_metrics'
 require 'rack'
 
 # A simple rack app for example.
 class RackApp
   def self.call _env
-    [200, { 'Content-Type' => 'text/html' }, ['Hi!']]
+    [200, { some_header: 'a value' }, ['Hi!']]
   end
 end
 
 app =
   Rack::Builder.app do
-    use(MetricsReporter)
+    use(RackMiddlewareMetrics::Reporter)
     run(RackApp)
   end
