+++ lib/rack_middleware_metrics.rb
@@ -1,6 +1,6 @@
 require "rack_middleware_metrics/version"
+require "rack_middleware_metrics/reporter"
 
 module RackMiddlewareMetrics
   class Error < StandardError; end
-  # Your code goes here...
 end
